ecs_container_with_sidecar/app/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-instrumentation-fastapi
opentelemetry-sdk 
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-requests
